On branch Homing-Movement Changes to be committed: 	new file:   TeamCode/src/main/java/org/firstinspires/ftc/teamcode/start_adb.sh 	new file:   TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vegamind/Gestell.java 	modified:   TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vegamind/Lifter.java 	modified:   TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vegamind/PrimaryTeleop.java
"PrimaryTeleop" is still the main TeleOp for the robot. Changed the movement to be a git more "standard" (Thanks "dpnoob" on discord).
Added a Homing sequence to the lifters to zero out the position and ensure proper placement of the lifters.
Added a routine that checks for motor blockages and reports them back to the operator via a Telemetry message (this needs work, sometimes fires incorrectly).